python/eks-basic/requirements.txt
aws-cdk-lib==2.236.0
aws-cdk.aws-eks-v2-alpha==2.236.0a0
aws-cdk.lambda-layer-kubectl-v34==2.0.0
constructs>=10.0.0,<11.0.0
